--- backend/dbscript.sql
@@ -1,11 +1,0 @@
--- -- For test puporposes, delete these queries.
--- UPDATE observaciones
--- SET impreso = 1
--- WHERE impreso = 0 AND id_registro = '1,137,01';
-
--- UPDATE observaciones
--- SET impreso = 0
--- WHERE impreso = 1 AND id_registro = '1,137,02';
-
--- DELETE FROM observaciones
--- WHERE id_registro = '1,137,01';
