+++ backend/inventario/_utils/__init__.py
@@ -1,0 +1,1 @@
+from inventario._utils.file_utils    import handle_uploaded_file
