+++ backend/inventario/_utils/file_utils.py
@@ -1,0 +1,91 @@
+import os
+from sgica.settings          import MEDIA_ROOT
+
+def count_files_in_directory(directory):
+    count = 1
+    for _, _, files in os.walk(directory):
+        count += len(files)
+    return count
+
+def handle_file_directories(doc_type:str | None = None,
+                            folder_name:str | None = None) -> list:
+    """
+        It creates a directory for the file in case that is a new product otherwise
+        it returns the path for the specified product
+    """
+    if doc_type is None:
+        absolute_path:str = f"{MEDIA_ROOT}/uploads/actas/"
+        relative_path:str = f"uploads/actas/"
+
+    if doc_type == "ubicacion_img":
+        absolute_path:str = f"{MEDIA_ROOT}/uploads/ubicaciones/{folder_name}/"
+        relative_path:str = f"uploads/ubicaciones/{folder_name}/"
+ 
+    if os.path.exists(absolute_path):
+        return [relative_path, absolute_path]
+
+    os.makedirs(absolute_path)
+    return [relative_path, absolute_path]
+
+
+def handle_uploaded_file(files: list,
+                         doc_type:str | None = None,
+                         **kwargs) -> str:
+    """
+        saves the file following this structure:
+        uploads/{model pk}/{file_name.ext}
+    """
+
+    nombre_oficial:str = kwargs.get("nombre_oficial", "")
+    nombre_oficial = nombre_oficial.replace(" ", "_") if nombre_oficial != "" else ""
+       
+    paths_list = handle_file_directories(doc_type = doc_type,
+                                         folder_name = nombre_oficial)
+    absolute_path = paths_list[1] 
+    relative_path = paths_list[0]
+    
+    if doc_type is None: 
+        for file in files:
+            relative_path = os.path.join(relative_path, str(file))
+            path_to_write = os.path.join(absolute_path, str(file)) 
+            with open(path_to_write, 'wb') as destination: 
+                for chunk in file.chunks():
+                    destination.write(chunk)
+
+    if doc_type == "ubicacion_img":
+        img_name = "img"
+        count = count_files_in_directory(directory = absolute_path)
+
+        for file in files:
+            ext = str(file).split(".")[1]  
+            path_to_write = os.path.join(absolute_path,
+                                         f"{img_name}{count}.{ext}")  
+            
+            with open(path_to_write, 'wb') as destination: 
+                for chunk in file.chunks():
+                    destination.write(chunk)
+            
+            count+=1 
+
+    return relative_path
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
