--- backend/my_name_is_scripty.py
@@ -1,8 +1,0 @@
-# HELLO GUYS MY NAME IS SCRIPTY AND I WILL DO EVERYTHING FOR YOU :)
-import requests
-import pandas
-import os
-import platform
-
-
-system_platform = platform.system()
